# libexpr: Add `EvalProfiler` and use it for `FunctionCallTrace`