<!--

IMPORTANT

Nix is a non-trivial project, so for your contribution to be successful,
it really is important to follow the contributing guidelines:

https://github.com/NixOS/nix/blob/master/CONTRIBUTING.md

Even if you've contributed to open source before, take a moment to read it,
so you understand the process and the expectations.

- what information to include in commit messages
- proper attribution
- volunteering contributions effectively
- how to get help and our review process.

-->

## Motivation

Having and extensible framework for evaluation profiling is much needed if we want to get flamegraph/tracy-based profilers https://github.com/NixOS/nix/pull/9967, https://github.com/NixOS/nix/pull/11373. This patch adds a `EvalProfiler` interface and `MultiEvalProfiler`, which support `preFunctionCallHook` and `postFunctionCallHook` that get call on entering and exiting `callFunction`. The actual hook invocation is guarded behind a simple cached boolean check with an `[[unlikely]]` annotation. This machinery is enough to reimplement existing `FunctionCallTrace`.

I hope this patch can pave the way forward for https://github.com/NixOS/nix/pull/9967, https://github.com/NixOS/nix/pull/11373 in an extensible and well-architected manner. Since the actual profiler implementation is hidden behind an interface, multiple profilers can be easily supported, so flamegraph/tracy can hopefully co-exist behind a `--eval-profiler` flag. (Current setting `trace-function-calls` can also be absorbed into that hypothetical feature).

(Inlined message from the second commit, which has non-tracing performance measurements)

Note that branches when the hook gets called are marked with `[[unlikely]]`
as a hint to the compiler that this is not a hot path. For non-tracing
evaluation this should be a 100% predictable branch, so the performance
cost is nonexistent.

Some measurements to support this point:

```
nix build .#nix-cli
nix build github:nixos/nix/d692729759e4e370361cc5105fbeb0e33137ca9e#nix-cli --out-link before
```

(Before)

```
$ taskset -c 2,3 hyperfine "GC_INITIAL_HEAP_SIZE=16g before/bin/nix eval nixpkgs#gnome --no-eval-cache" --warmup 4
Benchmark 1: GC_INITIAL_HEAP_SIZE=16g before/bin/nix eval nixpkgs#gnome --no-eval-cache
  Time (mean ± σ):      2.517 s ±  0.032 s    [User: 1.464 s, System: 0.476 s]
  Range (min … max):    2.464 s …  2.557 s    10 runs
```

(After)

```
$ taskset -c 2,3 hyperfine "GC_INITIAL_HEAP_SIZE=16g result/bin/nix eval nixpkgs#gnome --no-eval-cache" --warmup 4
Benchmark 1: GC_INITIAL_HEAP_SIZE=16g result/bin/nix eval nixpkgs#gnome --no-eval-cache
  Time (mean ± σ):      2.499 s ±  0.022 s    [User: 1.448 s, System: 0.478 s]
  Range (min … max):    2.472 s …  2.537 s    10 runs
```

<!-- Briefly explain what the change is about and why it is desirable. -->

## Context

https://github.com/NixOS/nix/pull/9967
https://github.com/NixOS/nix/pull/11373

<!-- Provide context. Reference open issues if available. -->

<!-- Non-trivial change: Briefly outline the implementation strategy. -->

<!-- Invasive change: Discuss alternative designs or approaches you considered. -->

<!-- Large change: Provide instructions to reviewers how to read the diff. -->

---

Add :+1: to [pull requests you find important](https://github.com/NixOS/nix/pulls?q=is%3Aopen+sort%3Areactions-%2B1-desc).

The Nix maintainer team uses a [GitHub project board](https://github.com/orgs/NixOS/projects/19) to [schedule and track reviews](https://github.com/NixOS/nix/tree/master/maintainers#project-board-protocol).